fix(wallet): allow precise ISK transfers by switching AlterBalance to double
Reworked Character::AlterBalance and Client::AddBalance to use double precision
instead of float, fixing a critical bug where small ISK transfers (e.g., 2,000 ISK)
were silently skipped due to float rounding.

- Changed AlterBalance(float) to AlterBalance(double)
- Updated AddBalance and all related calls to pass double
- Introduced precision-safe check using std::abs(amount) < 0.00001
- Added debug logging to detect skipped micro-transactions

This completes the fix for ISK not being deducted or added correctly
when processing low-value market orders.

+++ src/eve-server/character/Character.h
@@ -220,7 +220,7 @@ public:
     /**
      * Primary public interface:
      */
-    bool            AlterBalance(float amount, uint8 type);
+    bool            AlterBalance(double amount, uint8 type);  // ---sellorder update
     void            SetLocation(uint32 stationID, SystemData& data);
     void            JoinCorporation(const CorpData& data);
     void            SetDescription(const char *newDescription);
